requirements.txt
absl-py==2.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.121.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
joblib==1.5.2
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pip==21.2.4
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
setuptools==58.0.4
six==1.17.0
sklearn==0.0
sniffio==1.3.1
starlette==0.49.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
zipp==3.23.0
